feat(sln): move objectquery extensions, update docs/build config, bump version to 1.3.0
