+++ global.json
@@ -1,7 +1,7 @@
 {
   "sdk": {
-    "version": "6.0.401",
-    "rollForward": "latestMajor",
+    "version": "6.0.403",
+    "rollForward": "latestPatch",
     "allowPrerelease": false
   }
 }
