+++ src/XperienceCommunity.QueryExtensions/XperienceCommunityConnectionHelperExtensions.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using CMS.Helpers;
-using XperienceCommunity.QueryExtensions.Objects;
+using XperienceCommunity.QueryExtensions.DataSets;
 
 namespace CMS.DataEngine
 {
@@ -82,44 +82,15 @@ namespace CMS.DataEngine
         }
 
         /// <summary>
-        /// Executes the current query and returns it as a DataSet.  Extension method to convert ExecuteReaderAsync's IDataReader into a DataSet.
-        /// </summary>
-        /// <typeparam name="TObject">The Object Type</typeparam>
-        /// <param name="baseQuery"></param>
-        /// <param name="commandBehavior">Command behavior for the reader.</param>
-        /// <param name="newConnection">If true, the reader will be executed using its own dedicated connection.</param>
-        /// <param name="cancellationToken">The cancellation instruction.</param>
-        /// <returns>Returns a task returning either the data set with one table.</returns>
-        public static async Task<DataSet> ExecuteAsync<TObject>(this ObjectQuery<TObject> baseQuery, CommandBehavior commandBehavior = CommandBehavior.Default, bool newConnection = false, CancellationToken? cancellationToken = null) where TObject : BaseInfo, new()
-        {
-            var reader = await baseQuery.ExecuteReaderAsync(commandBehavior, newConnection, cancellationToken);
-            return DataReaderToDataSet(reader);
-        }
-
-        /// <summary>
-        /// Executes the current query and returns it as a DataSet.  Extension method to convert ExecuteReaderAsync's IDataReader into a DataSet.
-        /// </summary>
-        /// <param name="baseQuery"></param>
-        /// <param name="commandBehavior">Command behavior for the reader.</param>
-        /// <param name="newConnection">If true, the reader will be executed using its own dedicated connection.</param>
-        /// <param name="cancellationToken">The cancellation instruction.</param>
-        /// <returns>Returns a task returning either the data set with one table.</returns>
-        public static async Task<DataSet> ExecuteAsync(this ObjectQuery baseQuery, CommandBehavior commandBehavior = CommandBehavior.Default, bool newConnection = false, CancellationToken? cancellationToken = null)
-        {
-            var reader = await baseQuery.ExecuteReaderAsync(commandBehavior, newConnection, cancellationToken);
-            return DataReaderToDataSet(reader);
-        }
-
-        /// <summary>
         /// Converts a DbDataReader to a DataSet, handles multiple tables in return result.
         /// </summary>
         /// <param name="reader"></param>
         /// <returns></returns>
         private static DataSet DataReaderToDataSet(DbDataReader reader)
         {
-            if (reader == null)
+            if (reader is null)
             {
-                return EmptyDataSet();
+                return new DataSet().AddEmptyTable();
             }
 
             var ds = new DataSet();
@@ -131,38 +102,6 @@ namespace CMS.DataEngine
                 ds.Tables.Add(table);
             } while (!reader.IsClosed);
 
-            return ds;
-        }
-
-        /// <summary>
-        /// Converts a DbDataReader to a DataSet, handles multiple tables in return result.
-        /// </summary>
-        /// <param name="reader"></param>
-        /// <returns></returns>
-        private static DataSet DataReaderToDataSet(IDataReader reader)
-        {
-            if (reader == null)
-            {
-                return EmptyDataSet();
-            }
-
-            var ds = new DataSet();
-            // read each data result into a datatable
-            do
-            {
-                var table = new DataTable();
-                table.Load(reader);
-                ds.Tables.Add(table);
-            } while (!reader.IsClosed);
-
-            return ds;
-        }
-
-        private static DataSet EmptyDataSet()
-        {
-            var table = new DataTable();
-            var ds = new DataSet();
-            ds.Tables.Add(table);
             return ds;
         }
 
@@ -178,21 +117,23 @@ namespace CMS.DataEngine
             CacheHelper.CacheAsync(
                 async cs =>
                 {
-                    var query = await new ObjectQuery<QueryInfo>()
+                    var results = await new ObjectQuery<QueryInfo>()
                         .Where($"ClassID in (Select top 1 CMS_Class.ClassID from CMS_Class where ClassName = '{SqlHelper.EscapeQuotes(className)}')")
                         .WhereEquals(nameof(QueryInfo.QueryName), queryCodeName)
-                        .FirstOrDefaultAsync(token);
+                        .GetEnumerableTypedResultAsync(cancellationToken: token);
 
-                    if (query is null)
+                    var result = results.FirstOrDefault();
+
+                    if (result is null)
                     {
                         cs.Cached = false;
 
-                        return query;
+                        return result;
                     }
 
-                    cs.GetCacheDependency = () => CacheHelper.GetCacheDependency($"{QueryInfo.OBJECT_TYPE}|byid|{query.QueryID}");
+                    cs.GetCacheDependency = () => CacheHelper.GetCacheDependency($"{QueryInfo.OBJECT_TYPE}|byid|{result.QueryID}");
 
-                    return query;
+                    return result;
                 }
                 , new CacheSettings(Math.Clamp(cacheLengthMinutes, 1, int.MaxValue), nameof(GetCachedQueryAsync), className, queryCodeName));
     }
